Merge pyReDi project into pySDC
pyReDi (as the name already suggests) provides a framework for the
modeling, spatio-temporal discretization and numerical treatment of IBVPs of
reaction-diffusion systems. Its heart is the nonlinear FAS-MG solver specialized
in treating the systems arising in fully-implicit timestepping schemes.
This commit is to be considered more as a functional 'co-existence' than a
full marriage of the two projects. Serious restructuring for marriage lies ahead.